+++ core/DeviceConfig.php
@@ -435,7 +435,7 @@ abstract class DeviceConfig extends \core\common\Entity
         }
         // Okay, no custom suffix. 
         // Use the configured inst name and apply shortening heuristics
-        $lang_pointer = \config\Main::LANGUAGES[$this->languageInstance->getLang()]['latin_based'] == TRUE ? 0 : 1;
+        $lang_pointer = \config\Master::LANGUAGES[$this->languageInstance->getLang()]['latin_based'] == TRUE ? 0 : 1;
         $this->loggerInstance->debug(5, "getInstallerBasename1:" . $this->attributes['general:instname'][$lang_pointer] . "\n");
         $inst = $this->customTranslit($this->attributes['general:instname'][$lang_pointer]);
         $this->loggerInstance->debug(4, "getInstallerBasename2:$inst\n");